src/main/java/org/application/gameshelfapp/removevideogame/RemoveVideogameController.java
package org.application.gameshelfapp.removevideogame;

import org.application.gameshelfapp.buyvideogames.bean.VideogameBean;
import org.application.gameshelfapp.buyvideogames.dao.ItemDAO;
import org.application.gameshelfapp.buyvideogames.entities.Filters;
import org.application.gameshelfapp.buyvideogames.entities.Videogame;
import org.application.gameshelfapp.buyvideogames.exception.GameSoldOutException;
import org.application.gameshelfapp.login.dao.PersistencyAbstractFactory;
import org.application.gameshelfapp.login.exception.CheckFailedException;
import org.application.gameshelfapp.login.exception.PersistencyErrorException;
import org.application.gameshelfapp.sellvideogames.exception.NoGameInCatalogueException;

public class RemoveVideogameController {
    public void removeGameFromCatalogue(VideogameBean videogameBean) throws PersistencyErrorException, NoGameInCatalogueException, GameSoldOutException, CheckFailedException {
        ItemDAO itemDAO = PersistencyAbstractFactory.getFactory().createItemDAO();
        Filters filters = new Filters( videogameBean.getPlatformBean(), videogameBean.getCategoryBean());
        filters.setName(videogameBean.getName());

        Videogame game = itemDAO.getVideogame(filters);
        game.removeVideogame(videogameBean.getCopiesBean());
        itemDAO.removeGameForSale(game);
    }
}
